+++ braid/src/examples/animals.rs
@@ -1,8 +1,6 @@
 //! Utilities for the animals example
-use std::convert::TryInto;
-
 use crate::examples::IndexConversionError;
-use crate::{ColumnIndex, NameOrIndex, RowIndex, TableIndex};
+use std::convert::TryInto;
 
 /// Row names for the animals data set
 #[repr(usize)]
@@ -63,12 +61,6 @@ pub enum Row {
 impl From<Row> for usize {
     fn from(row: Row) -> Self {
         row as Self
-    }
-}
-
-impl From<Row> for RowIndex {
-    fn from(row: Row) -> Self {
-        Self(NameOrIndex::Index(row.into()))
     }
 }
 
@@ -240,12 +232,6 @@ impl Row {
 impl From<Column> for usize {
     fn from(col: Column) -> Self {
         col as Self
-    }
-}
-
-impl From<Column> for ColumnIndex {
-    fn from(col: Column) -> Self {
-        Self(NameOrIndex::Index(col.into()))
     }
 }