+++ braid/src/interface/engine/error.rs
@@ -5,8 +5,6 @@ use braid_cc::feature::FType;
 use braid_codebook::CodebookError;
 use thiserror::Error;
 
-use crate::TableIndex;
-
 /// Errors that can arise when parsing data for an Engine
 #[derive(Debug, Error)]
 pub enum DataParseError {
@@ -146,19 +144,18 @@ pub enum InsertDataError {
         `{col}`"
     )]
     NonFiniteContinuousValue { col: String, value: f64 },
-    #[error(
-        "Supplied an usize index variant ({0}) for a row that does not exist. \
-         Non-existent indices must be given by name."
-    )]
-    UsizeRowIndexOutOfBounds(usize),
-    #[error(
-        "Supplied an usize index variant ({0}) for a column that does not \
-         exist. Non-existent indices must be given by name."
-    )]
-    UsizeColumnIndexOutOfBounds(usize),
+    #[error("Row index error: {0}")]
+    RowIndex(IndexError),
+    #[error("Column index error: {0}")]
+    ColumnIndex(IndexError),
     /// An placeholder error variant used when chaining `ok_or` with `map_or`
     #[error("How can you extract what is unreachable?")]
     Unreachable,
+    #[error(
+        "The column with usize index '{0}' appears to be new, but new columns \
+        must be given string names"
+    )]
+    IntergerIndexNewColumn(String),
 }
 
 /// Errors that can arise when removing data from the engine