+++ braid/src/interface/engine/mod.rs
@@ -305,7 +305,7 @@ impl Engine {
     /// let starting_rows = engine.n_rows();
     ///
     /// let rows = vec![
-    ///     Row {
+    ///     Row::<&str, &str> {
     ///         row_ix: "pegasus".into(),
     ///         values: vec![
     ///             Value {
@@ -351,7 +351,7 @@ impl Engine {
     /// use braid_codebook::{ColMetadataList, ColMetadata, ColType};
     /// use braid_stats::prior::csd::CsdHyper;
     ///
-    /// let rows: Vec<Row> = vec![
+    /// let rows: Vec<Row<&str, &str>> = vec![
     ///     ("bat", vec![("drinks+blood", Datum::Categorical(1))]).into(),
     ///     ("beaver", vec![("drinks+blood", Datum::Categorical(0))]).into(),
     /// ];
@@ -400,7 +400,7 @@ impl Engine {
     /// use braid_codebook::{ColMetadataList, ColMetadata, ColType};
     /// use braid_stats::prior::csd::CsdHyper;
     ///
-    /// let rows: Vec<Row> = vec![
+    /// let rows: Vec<Row<&str, &str>> = vec![
     ///     ("bat", vec![
     ///             ("drinks+blood", Datum::Categorical(1)),
     ///     ]).into(),
@@ -471,7 +471,7 @@ impl Engine {
     /// // Turns out pigs are fierce.
     /// assert_eq!(x_before, Datum::Categorical(1));
     ///
-    /// let rows: Vec<Row> = vec![
+    /// let rows: Vec<Row<&str, &str>> = vec![
     ///     // Inserting a 2 into a binary column
     ///     ("pig", vec![("fierce", Datum::Categorical(2))]).into(),
     /// ];
@@ -528,7 +528,7 @@ impl Engine {
     ///     }
     /// };
     ///
-    /// let rows: Vec<Row> = vec![(
+    /// let rows: Vec<Row<&str, &str>> = vec![(
     ///     "Artemis (Advanced Data Relay and Technology Mission Satellite)",
     ///     vec![("Class_of_Orbit", Datum::Categorical(4))]
     /// ).into()];
@@ -542,14 +542,14 @@ impl Engine {
     ///
     /// assert!(result.is_ok());
     /// ```
-    pub fn insert_data(
+    pub fn insert_data<R: RowIndex, C: ColumnIndex>(
         &mut self,
-        rows: Vec<Row>,
+        rows: Vec<Row<R, C>>,
         new_metadata: Option<ColMetadataList>,
         suppl_metadata: Option<HashMap<String, ColMetadata>>,
         mode: WriteMode,
     ) -> Result<InsertDataActions, InsertDataError> {
-        use data::standardize_rows_for_insert;
+        // use data::standardize_rows_for_insert;
         // TODO: Lots of opportunity for optimization
         // TODO: Errors not caught
         // - user inserts missing data into new column so the column is all
@@ -560,7 +560,7 @@ impl Engine {
 
         // Convert the indices into usize if present and string/name if not
         // Error if the user has passed an usize index that is out of bounds
-        let rows = standardize_rows_for_insert(rows, &self.codebook)?;
+        // let rows = standardize_rows_for_insert(rows, &self.codebook)?;
 
         // Figure out the tasks required to insert these data, and convert all
         // String row/col indices into usize.
@@ -648,39 +648,33 @@ impl Engine {
     /// Remove a cell.
     /// ```rust
     /// # use braid::examples::Example;
-    /// use braid::examples::animals::{Row, Column};
-    /// use braid::{TableIndex, NameOrIndex, OracleT};
+    /// use braid::{TableIndex, OracleT};
     /// use braid_data::Datum;
-    ///
-    /// let horse: usize = Row::Horse.into();
-    /// let flys: usize = Column::Flys.into();
     ///
     /// let mut engine = Example::Animals.engine().unwrap();
     ///
-    /// assert_eq!(engine.datum(horse, flys).unwrap(), Datum::Categorical(0));
+    /// assert_eq!(engine.datum("horse", "flys").unwrap(), Datum::Categorical(0));
     ///
     /// // Row and Column implement Into<TableIndex>
-    /// engine.remove_data(vec![(horse, flys).into()]);
+    /// engine.remove_data(vec![("horse", "flys").into()]);
     ///
-    /// assert_eq!(engine.datum(horse, flys).unwrap(), Datum::Missing);
+    /// assert_eq!(engine.datum("horse", "flys").unwrap(), Datum::Missing);
     /// ```
     ///
     /// Remove a row and column.
     ///
     /// ```rust
     /// # use braid::examples::Example;
-    /// # use braid::examples::animals::{Row, Column};
-    /// # use braid::{TableIndex, NameOrIndex, OracleT, HasStates};
+    /// # use braid::{TableIndex, OracleT, HasStates};
     /// # use braid_data::Datum;
     /// let mut engine = Example::Animals.engine().unwrap();
     ///
     /// assert_eq!(engine.n_rows(), 50);
     /// assert_eq!(engine.n_cols(), 85);
     ///
-    /// // Row and Column implement Into<TableIndex>
     /// engine.remove_data(vec![
-    ///     Row::Horse.into(),
-    ///     Column::Flys.into(),
+    ///     TableIndex::Row("horse"),
+    ///     TableIndex::Column("flys"),
     /// ]);
     ///
     /// assert_eq!(engine.n_rows(), 49);
@@ -691,8 +685,7 @@ impl Engine {
     ///
     /// ```rust
     /// # use braid::examples::Example;
-    /// # use braid::examples::animals::{Row, Column};
-    /// # use braid::{TableIndex, NameOrIndex, OracleT, HasStates};
+    /// # use braid::{TableIndex, OracleT, HasStates};
     /// # use braid_data::Datum;
     /// let mut engine = Example::Animals.engine().unwrap();
     ///